GCC isn't clever enough to eliminate all dead stores. Besides, we don't have to store a copy of destroyer along with every object.

Besides, I missed a data structure that should be `std::hardware_destructive_interference_size` aligned.

When tested with our echo server (`example/rpc/server.cc`), this optimization reduces slow path of `Get`-ting fiber stacks (actually, moving them from central per-node pool) from ~2000ns down to ~900ns, `Put`-ting them from ~2000ns down to ~1000ns, respectively.